refactor: Dependencyをリネームしてmainに移動
